requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dataclasses==0.6
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-mongoengine==0.5.5
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djongo==1.3.7
dnspython==2.7.0
drf-yasg==1.21.10
inflection==0.5.1
kombu==5.5.1
mongoengine==0.24.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
PyJWT==2.9.0
pymongo==3.11.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
